backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
pydantic>=2.0.0
langchain>=0.0.300
langchain-openai>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0 